Remove all use of `skip` method.
Because reverse iteration in Memento was broken, it would not correctly
do less than or equal and would start off by one record, I was filtering
the first record using Memento `skip`. The previous commit fixes an
error that was based on that bug. This fixes an error in reasoning about
index collections. We cannot use Memento's `exclusive` cursor builder
method because we are always searching by key so we use the special
`MAX` value when exclusive to go past all entries for a particular key.

Closes #25.